requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
djangorestframework==3.15.2
idna==3.10
mysqlclient==2.2.7
pillow==11.1.0
python-dotenv==1.1.0
requests==2.32.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
urllib3==2.3.0
